refactor: remove AuthProvider component and implement Login component for improved authentication handling
