+++ src/pages/admin/index.astro
@@ -1,27 +1,28 @@
 ﻿---
 import Layout from '@/layouts/layout.astro'
-import AuthProvider from '@/components/AuthProvider.vue'
-import DashboardAnswers from '@/components/DashboardAnswers.vue'
 import SurveyList from '@/components/SurveyList.astro'
+import Login from '@/components/Login.vue'
 ---
 
 <Layout>
-  <AuthProvider client:load>
-    <h1>Inquéritos</h1>
+  <h1>Inquéritos</h1>
 
-    <div class="l-row l-row--start l-row--large">
-      <div class="l-stack l-stack--x-small">
-        <h2>Inquéritos passados</h2>
-        <SurveyList status="previous" />
-      </div>
-      <div class="l-stack l-stack--x-small">
-        <h2>Inquéritos a decorrer</h2>
-        <SurveyList status="current" />
-      </div>
-      <div class="l-stack l-stack--x-small">
-        <h2>Inquéritos futuros</h2>
-        <SurveyList status="future" />
-      </div>
+  <div class="is-visible-logged-out">
+    <Login client:visible />
+  </div>
+
+  <div class="l-row l-row--start l-row--large is-visible-logged-in">
+    <div class="l-stack l-stack--x-small">
+      <h2>Inquéritos passados</h2>
+      <SurveyList status="previous" server:defer />
     </div>
-  </AuthProvider>
+    <div class="l-stack l-stack--x-small">
+      <h2>Inquéritos a decorrer</h2>
+      <SurveyList status="current" server:defer />
+    </div>
+    <div class="l-stack l-stack--x-small">
+      <h2>Inquéritos futuros</h2>
+      <SurveyList status="future" server:defer />
+    </div>
+  </div>
 </Layout>
